refactor: #439 PR 1 - using seperate modules for grouping models instead of a single models.py file
