+++ .gitignore
@@ -38,3 +38,4 @@ coverage.xml
 
 # Media files
 media/
+private_files/
