+++ django_email_learning/models/enums/deactivation_reason.py
@@ -1,0 +1,8 @@
+from enum import StrEnum
+
+
+class DeactivationReason(StrEnum):
+    CANCELED = "canceled"
+    BLOCKED = "blocked"
+    FAILED = "failed"
+    INACTIVE = "inactive"
