+++ django_email_learning/models/enums/delivery_status.py
@@ -1,0 +1,9 @@
+from enum import StrEnum
+
+
+class DeliveryStatus(StrEnum):
+    SCHEDULED = "scheduled"
+    PROCESSING = "processing"
+    DELIVERED = "delivered"
+    CANCELED = "canceled"
+    BLOCKED = "blocked"
