+++ django_email_learning/models/enums/enrollment_status.py
@@ -1,0 +1,8 @@
+from enum import StrEnum
+
+
+class EnrollmentStatus(StrEnum):
+    UNVERIFIED = "unverified"
+    ACTIVE = "active"
+    COMPLETED = "completed"
+    DEACTIVATED = "deactivated"
