+++ tests/models/test_course.py
