+++ tests/models/test_course_instructor.py
